tests/WeKan.Application.UnitTests/Commands/DeleteActivity/DeleteActivityAuthorizerTests.cs
using Moq;
using System;
using System.Threading;
using System.Threading.Tasks;
using WeKan.Application.Commands.DeleteActivity;
using WeKan.Application.Common.Interfaces;
using WeKan.Domain.Boards;
using Xunit;

namespace WeKan.Application.UnitTests.Commands.DeleteActivity
{
    public class DeleteActivityAuthorizerTests
    {
        private readonly Mock<ICurrentUserPermissionService> _currentUserPermissionService;

        public DeleteActivityAuthorizerTests()
        {
            _currentUserPermissionService = new Mock<ICurrentUserPermissionService>();
        }

        [Fact]
        public void Ctor_ICurrentUserPermissionServiceNull_ThrowsArgumentNullException()
        {
            Assert.Throws<ArgumentNullException>(() => new DeleteActivityAuthorizer(null));
        }

        [Fact]
        public async Task Authorise_CallsHasPermissionForActivity_CanDeleteActivity()
        {
            var permission = BoardUserPermission.CAN_DELETE_ACTIVITY;
            var expectedResult = true;
            CancellationToken cancellationToken = new CancellationToken();
            _currentUserPermissionService
                .Setup(s => s.HasPermissionForActivity(
                    It.IsAny<int>(),
                    It.Is<BoardUserPermission>(p => p == permission),
                    cancellationToken))
                .ReturnsAsync(expectedResult);

            var authorizer = new DeleteActivityAuthorizer(_currentUserPermissionService.Object);

            var result = await authorizer.Authorise(new DeleteActivityCommand(1), cancellationToken);

            Assert.Equal(expectedResult, result);
            _currentUserPermissionService
                .Verify(s => s.HasPermissionForActivity(It.IsAny<int>(), It.Is<BoardUserPermission>(p => p == permission), cancellationToken), Times.Once);
        }
    }
}
